requirements.txt
cvzone
mediapipe
tensorflow
streamlit 
opencv-python 
pillow
`numpy